refactor(team): split broker_helpers.go into 3 cohesive files
Staff-reviewer feedback on the broker decomposition: broker_helpers.go
was a junk drawer wedging three unrelated concerns together. Split per
the review:

(1) broker_member_construction.go (105 LOC):
    - memberFromSpec (canonical officeMember constructor)
    - applyOfficeMemberDefaults (fill Name / Role / Expertise /
      Personality / PermissionMode when input is partial)
    - inferOfficeExpertise / inferOfficePersonality
    All feed the same pipeline: convert thin spec → fully-populated
    officeMember ready to persist.

(2) broker_channel_access.go (130 LOC):
    - reservedChannelSlugs (the security invariant)
    - canAccessChannelLocked (the policy decision)
    - channelHasMemberLocked / channelMemberEnabledLocked /
      enabledChannelMembersLocked (its data lookups)
    - ensureTaskOwnerChannelMembershipLocked (auto-promotion that
      shares the "every actor that posts/owns work must be a Member"
      invariant)
    The reservedChannelSlugs map and canAccessChannelLocked must stay
    in sync — colocating them here surfaces that coupling. The
    channel-create handler in broker_office_channels.go is the third
    coupled site and references reservedChannelSlugs by import.

(3) broker_text.go (131 LOC):
    - extractMentionedSlugs + mentionPattern
    - uniqueSlugs / normalizeStringList / containsString
    - humanizeSlug
    - defaultTeamChannelDescription
    - truncateSummary
    Pure text/list helpers; no broker state.

  - internal/team/broker_helpers.go: -326 LOC (deleted)
  - internal/team/broker_member_construction.go: +105 LOC (new)
  - internal/team/broker_channel_access.go: +130 LOC (new)
  - internal/team/broker_text.go: +131 LOC (new)

Net: +40 LOC across 3 files vs 1 file (extra header docstrings + the
"why these belong together" comment per file). Trade is intentional —
each file now reads as a themed bundle.

Tests pass.
